Avoid warnings due to implicit time conversions.

+++ src/api/mcfrelax.c
@@ -157,12 +157,12 @@ int glp_mincost_relax4(glp_graph *G, int v_rhs, int a_low, int a_cap,
             /* substitute x = x' + low, where 0 <= x' <= cap - low */
             csa.u[k] = (int)(cap - low);
             /* correct demands at endpoints of k-th arc */
-            if (overflow(csa.dfct[a->tail->i], +low))
+            if (overflow(csa.dfct[a->tail->i], (int) +low))
             {  ret = GLP_ERANGE;
                goto done;
             }
-            csa.dfct[a->tail->i] += low;
-            if (overflow(csa.dfct[a->head->i], -low))
+            csa.dfct[a->tail->i] += (int) low;
+            if (overflow(csa.dfct[a->head->i], (int) -low))
             {  ret = GLP_ERANGE;
                goto done;
             }
